# frontend/docs/doc2md_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ingestify API",
    "description": "\nAPI assíncrona para conversão de documentos para Markdown usando Docling\n\n## Autenticação\n\nEsta API suporta dois métodos de autenticação:\n\n### 1. Bearer Token (JWT)\n1. Registre um usuário em `/auth/register`\n2. Faça login em `/auth/login` para obter o token\n3. Clique no botão **\"Authorize\"** (cadeado 🔒) no topo desta página\n4. Cole o token no campo \"bearerAuth\" (sem prefixo \"Bearer\")\n5. Clique em \"Authorize\"\n\n### 2. API Key\n1. Faça login para obter um token JWT\n2. Crie uma API Key em `/api-keys/`\n3. Clique no botão **\"Authorize\"** (cadeado 🔒)\n4. Cole a API Key no campo \"apiKeyAuth\"\n5. Clique em \"Authorize\"\n\nApós autorizar, todos os endpoints protegidos usarão automaticamente suas credenciais.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/admin/stats": {
      "get": {
        "tags": [
          "Admin & Monitoring"
        ],
        "summary": "Get system statistics",
        "description": "Get comprehensive system statistics for monitoring dashboard\n\nReturns counts of jobs/pages by status, stuck jobs, etc.\nUseful for building admin dashboards and monitoring tools.",
        "operationId": "get_stats_admin_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Stats Admin Stats Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/admin/jobs/stuck": {
      "get": {
        "tags": [
          "Admin & Monitoring"
        ],
        "summary": "List stuck jobs",
        "description": "List all jobs currently stuck in processing state\n\nArgs:\n    threshold_minutes: Override default threshold (from config if not specified)\n    limit: Maximum number of jobs to return\n\nReturns:\n    List of stuck jobs with details",
        "operationId": "list_stuck_jobs_admin_jobs_stuck_get",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "threshold_minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Threshold Minutes"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Stuck Jobs Admin Jobs Stuck Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/jobs/recover-stuck": {
      "post": {
        "tags": [
          "Admin & Monitoring"
        ],
        "summary": "Manually trigger stuck job recovery",
        "description": "Manually trigger the stuck job detection and recovery process\n\nThis runs the same logic as the periodic monitoring task,\nbut can be triggered on-demand by administrators.\n\nArgs:\n    threshold_minutes: Override default threshold (from config if not specified)\n\nReturns:\n    Number of jobs/pages marked as failed",
        "operationId": "recover_stuck_jobs_admin_jobs_recover_stuck_post",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "threshold_minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Threshold Minutes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Recover Stuck Jobs Admin Jobs Recover Stuck Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/jobs/{job_id}/retry-all-failed": {
      "post": {
        "tags": [
          "Admin & Monitoring"
        ],
        "summary": "Bulk retry all failed pages of a job",
        "description": "Retry all failed pages of a specific job\n\nThis is useful when multiple pages failed due to temporary issues\nand you want to retry them all at once instead of individually.\n\nArgs:\n    job_id: The job ID whose failed pages should be retried\n\nReturns:\n    Number of pages queued for retry",
        "operationId": "retry_all_failed_pages_admin_jobs__job_id__retry_all_failed_post",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Retry All Failed Pages Admin Jobs  Job Id  Retry All Failed Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/cleanup": {
      "post": {
        "tags": [
          "Admin & Monitoring"
        ],
        "summary": "Manually trigger cleanup of old jobs",
        "description": "Manually trigger cleanup of old completed/failed jobs from Redis\n\nArgs:\n    days_old: Override default days threshold (from config if not specified)\n\nReturns:\n    Number of jobs cleaned up",
        "operationId": "trigger_cleanup_admin_cleanup_post",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "days_old",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Days Old"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trigger Cleanup Admin Cleanup Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/broker/unacked": {
      "get": {
        "tags": [
          "Admin & Monitoring"
        ],
        "summary": "List unacknowledged broker messages",
        "description": "Messages the Celery broker holds as delivered but not yet acknowledged.\n\nEach one is either a task running now or one whose worker died - the latter only\nreturns to its queue after CELERY_VISIBILITY_TIMEOUT_SECONDS (hours). `orphans` is\nthe latest check by workers.monitoring.check_broker_unacked: messages no live\nworker held on two consecutive checks. Requeue one with\nPOST /admin/broker/unacked/{delivery_tag}/requeue.",
        "operationId": "list_broker_unacked_admin_broker_unacked_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Broker Unacked Admin Broker Unacked Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtido via /auth/login"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key criada via /api-keys/"
      }
    }
  }
}